refactor: create api folder
